requirements.txt
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.3
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.0
nltk==3.7
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2022.9.13
six==1.16.0
tqdm==4.64.1
